Upgrade 30.07

+++ src/main/java/ua/foxminded/javaspring/school_console_app/model/Group.java
@@ -4,19 +4,19 @@ import java.util.Objects;
 
 public class Group {
 
-  private int id;
+  private long id;
   private String name;
 
-  public Group(int id, String name) {
+  public Group(long id, String name) {
     this.id = id;
     this.name = name;
   }
 
-  public int getId() {
+  public long getId() {
     return id;
   }
 
-  public void setId(int id) {
+  public void setId(long id) {
     this.id = id;
   }
